Added info about nmea data to README
I thought it might be useful to add an explanation of NMEA 0183 and the different formats in which the data is transmitted to the microcontroller from the GPS. Only caveat is, I essentially plagiarized all of it. I did credit and link to where I got it from, but I'm not sure about the proper way to cite it is.